% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epifish.R
\name{drawLegend2}
\alias{drawLegend2}
\title{Add a cluster legend to a fishplot}
\usage{
drawLegend2(
  fish,
  xpos = 0,
  ypos = -5,
  nrow = NULL,
  cex = 1,
  widthratio = 1.5,
  xsp = 0.5
)
}
\arguments{
\item{fish}{fishplot object}

\item{xpos}{distance from left edge of fishplot}

\item{ypos}{distance from bottom edge of fishplot}

\item{nrow}{number of rows to split clusters into}

\item{cex}{size of text}

\item{widthratio}{width of columns relative to longest legend entry}

\item{xsp}{horizontal spacing factor}
}
\description{
A modified version of the fishplot::drawLegend() function to allow more customisation of
the legend formatting, especially horizontal spacing when legend entries are long.
Refer to https://github.com/chrisamiller/fishplot/blob/master/R/draw.R
}
