% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek_calculations.R
\name{get_epi_week_span_date}
\alias{get_epi_week_span_date}
\title{Produce a human-readable start and end date for the epi week}
\usage{
get_epi_week_span_date(
  cdate,
  return_end = FALSE,
  add_year = FALSE,
  first_month = TRUE,
  date_format = "ymd"
)
}
\arguments{
\item{cdate}{character string in specified date format (eg 2020/01/25, 2020.1.25, 20-01-05 or the like)}

\item{return_end}{TRUE/FALSE whether to just return the end day}

\item{add_year}{TRUE/FALSE whether to include the year}

\item{first_month}{TRUE/FALSE whether to include the first month in the text}

\item{date_format}{character date format: "ymd" "dmy" or "mdy" (default is "ymd")}
}
\value{
character string
}
\description{
Given a case date, returns a text description of the epi week it represents
(eg "29 Mar - 4 Apr"). If `return_end` is `TRUE` just returns the end date
(eg "4 Apr"). If `add_year` is `TRUE` appends the year (eg "4 Apr 2020"). If
`first_month` is FALSE will skip the first month entry (eg "29-4 Apr")
}
\details{
A CDC-format "epi week" starts on Sunday and ends on Saturday. This function takes a case
date, and returns a human-friendly description of that week (useful for labelling plots).
}
\examples{
get_epi_week_span_date( "2020/01/25")
get_epi_week_span_date( "25.1.20", return_end=TRUE, add_year=TRUE, date_format="dmy")
get_epi_week_span_date( "1-25-2020", first_month=FALSE, add_year=TRUE, date_format="mdy")

}
