% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek_calculations.R
\name{get_epimonth}
\alias{get_epimonth}
\title{Identify epidemic month number for a given date}
\usage{
get_epimonth(cdate, start_date, date_format = "ymd")
}
\arguments{
\item{cdate}{character date of interest in chosen format}

\item{start_date}{character date of epidemic start}

\item{date_format}{character date format: "ymd" "dmy" or "mdy" (default is "ymd")}
}
\value{
integer number of months since the epidemic start
}
\description{
Given a case date and epidemic start date, returns the epidemic month it represents
}
\details{
This function takes a case date, and date to start counting on (eg the first case of an
epidemic or the first day of a year), and returns the month that case ocurred on.
}
\examples{
get_epimonth(cdate="2020/4/1", start_date="2020/1/25")
get_epimonth("25/1/2020", "25/1/2020", date_format="dmy")
get_epimonth("01-02-2020", "01-01-2019", date_format="mdy")

}
