% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek_calculations.R
\name{get_epiweek}
\alias{get_epiweek}
\title{Identify epi week number for a given date}
\usage{
get_epiweek(cdate, start_date, date_format = "ymd")
}
\arguments{
\item{cdate}{character date of interest in chosen format}

\item{start_date}{character date of epidemic start}

\item{date_format}{date format: "ymd" "dmy" or "mdy" (default is "ymd")}
}
\value{
integer number of epi weeks since the epidemic start
}
\description{
Given a case date and epidemic start date, returns the epi week it represents.
}
\details{
A CDC-format "epi week" starts on Sunday and ends on Saturday. This function takes a case
date, and date to start counting on (eg the first case of an epidemic or the first day
of a year), and returns the epi week that case ocurred on.
}
\examples{
get_epiweek(cdate="2020/4/1", start_date="2020/1/25")
get_epiweek("25/1/2020", "25/1/2020", date_format="dmy")
get_epiweek("01-01-2020", "01-01-2019", date_format="mdy")

}
