% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epifish.R
\name{make_parent_list}
\alias{make_parent_list}
\title{Create fishplot parent/child position vector}
\usage{
make_parent_list(parent_df, fishplot_names)
}
\arguments{
\item{parent_df}{data frame with columns "parent" and "child"}

\item{fishplot_names}{the fishplot table's column (cluster) names}
}
\value{
list of form ("child_cluster_name"= parent_column_number)
}
\description{
Fishplots require an input vector/list that matches each cluster in the fishplot table to its
parent's position in the table (or 0 if it is an independent cluster). This function
generates that vector for you, given a data frame naming the parent/child pairings.
}
\details{
Takes a dataframe describing parent-child relationships and a vector of fishplot cluster names
and generates the named child->parent position list needed for a fishplot and for the
`pad_parents()` function. An example parent data frame file is included with the epifish package.
Clusters with `NA` in the parent column will be considered independently derived.
}
