% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epifish.R
\name{pad_matrix}
\alias{pad_matrix}
\title{Pad "temporarily disappearing" clusters with a small value}
\usage{
pad_matrix(df, padval = 0.001)
}
\arguments{
\item{df}{data frame: fishplot table, after normalisation but before matrix transformation}

\item{padval}{numeric: the count value to pad with (defualt 0.001)}
}
\value{
list of form ("clustername"= "colour")
}
\description{
Pad "temporarily disappearing" clusters with a small value
}
\details{
Add tiny value to timepoints in the vector where cluster counts temporarily go
to zero before the cluster disappears, in order to fishplot rules where no cluster
can go to zero and then come back later.
Needed because setting "fix.missing.clones=TRUE" in the fishplot funciton doesn't actually work.
}
