% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epifish.R
\name{pad_parents}
\alias{pad_parents}
\title{Adjust counts for fishplot parent/child relationship rules}
\usage{
pad_parents(fish_table, parents)
}
\arguments{
\item{fish_table}{count table/matrix}

\item{parents}{named list of parent/child positions in the table (output of make_parent_list())}
}
\value{
fish_table with padded parent values
}
\description{
Add child counts to parent counts for all clusters with a parent/child relationship.
}
\details{
Fishplots require parent cluster counts to include the counts of all their
children as well (no parent can ever have a lower count than its children). This function
pads counts for parents with the counts for all their children to fulfil this rule.
NOTE to handle nesting this function relies on the fact that children must arise after parents
and that `parents` is a time-sorted list. If your data contains complex nested parent/child
relationships, check the output messages to ensure the additions occurred in the correct order.
}
