% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epifish.R
\name{set_fish_colours}
\alias{set_fish_colours}
\title{Create fishplot cluster colour vector}
\usage{
set_fish_colours(colour_df, fishplot_names)
}
\arguments{
\item{colour_df}{data frame with columns "cluster" and "colour"}

\item{fishplot_names}{the fishplot table's column (cluster) names}
}
\value{
list of form ("clustername"= "colour")
}
\description{
Converts an input cluster-colour pairing table to a named list in the proper order
for the fishplot, to enable easy use of custom colour palettes.
}
\details{
Takes a dataframe describing cluster-colour pairings, and generates a named list to
use when generating the fishplot. Colours can be described as R named colours
(https://www.nceas.ucsb.edu/sites/default/files/2020-04/colorPaletteCheatsheet.pdf) or hexadecimal codes.
An example input table is included in the epifish package.
}
