% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epifish.R
\name{build_epifish}
\alias{build_epifish}
\title{Build an epifish object for plotting}
\usage{
build_epifish(
  sample_df,
  parent_df = NULL,
  colour_df = NULL,
  min_cluster_size = 1,
  timepoint_labels = FALSE,
  label_clusters = TRUE,
  label_pos = 2,
  label_cex = 0.7,
  label_col = "black",
  label_angle = 0,
  label_offset = 0.2
)
}
\arguments{
\item{sample_df}{input sample data with character columns `cluster_id`, `timepoint`, and optionally `timepoint_label`}

\item{parent_df}{[optional] data frame with character columns `parent` and `child` describing cluster relationships; leave `NA` for independent clusters}

\item{colour_df}{[optional] data frame with character columns `cluster` and `colour`}

\item{min_cluster_size}{[optional] numeric whether to collapse all clusters < a given size together (default 1 to s}

\item{timepoint_labels}{[optional] TRUE/FALSE whether to look for a `timepoint_label` column in `sample_df` to use as the timepoint labels}
}
\value{
A list with the components:
 \describe{
  \item{fish}{epi-fishplot object to pass to fishplot::fishPlot()}
  \item{timepoint_counts}{summary table of number of samples per cluster per timepoint}
  \item{timepoint_sums}{summary table of number of samples per timepoint}
  \item{cluster_sums}{summary table of total number of samples per cluster}
  \item{timepoints}{vector of timepoints used}
  \item{timepoint_labels}{vector of the names of timepoints assigned in the plot}
  \item{parents}{named list matching child clusters to their parent's position in the matrix (0 means cluster is independent)}
  \item{raw_table}{initial table of counts per cluster per timepoint, before padding and normalisation}
  \item{fish_table}{normalised and parent-padded table for the epi-fishplot}
  \item{fish_matrix}{final transformed matrix used to make the epifish object}
}
}
\description{
Convert sample, cluster, and colour data frames into an epi-fishplot and a list of assorted useful summaries
}
\details{
Takes a data frame of epi sample data (one sample per row, containing columns `cluster_id` and `timepoint`).
From this, generates a fishplot-format normalised matrix of counts per cluster at each timepoint, a
fishplot object generated with this matrix, and assorted useful summary data.

Optionally, can take data frames of parent/child pairings to use for calculating the matrix, a data frame
with a custom cluster colour scheme, and a `timepoint_label` column in the sample data to generate custom
timepoint labels for the fishplot.
}
