% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiweek_calculations.R
\name{get_epiweek}
\alias{get_epiweek}
\title{Identify epidemic week number for a given date}
\usage{
get_epiweek(cdate, start_date, date_format = "ymd", week_start = 7)
}
\arguments{
\item{cdate}{character date of interest in chosen format}

\item{start_date}{character date of epidemic start}

\item{date_format}{character date format: "ymd" "dmy" or "mdy" (default is "ymd")}

\item{week_start}{integer day that weeks start on (1=Monday, 6=Saturday, 7=Sunday)}
}
\value{
integer number of epi weeks since the epidemic start
}
\description{
Given a case date and epidemic start date, returns the epidemic week it represents
}
\details{
This function takes a case date, and date to start counting on (eg the first case of an
epidemic or the first day of a year), and returns the week that case ocurred on.

By default this function uses a CDC-format "epi week" which starts on Sunday and ends on
Saturday. If you want your week to start on a different day, use the `week_start` variable.
#'
}
\examples{
get_epiweek(cdate="2020/4/1", start_date="2020/1/25", week_start=7)
get_epiweek("25/1/2020", "25/1/2020", date_format="dmy", week_start=7)
get_epiweek("01-01-2020", "01-01-2019", date_format="mdy", week_start=1)

}
